Link provider logs and document logs by UUID (#113)
Until now we had a foreign key relation pointing from provider logs to
document logs. But is hard to enforce it given that document logs will
in some ocassiones be generated after provider logs. The idea here is to
use a common UUID to fetch logs by `documentLogUuid` this way is easier
to sync paying the price of not having a databse enforced foreign key.