Use node-fetch on javascript SDK (#442)
We realized that the sdk is run on client's server and we can't control
the nodejs version or if the have `fetch` polyfill. Also one user saw
our sdk failing with TLSSocket.onSocketEnd TypeError terminated
other.side closed. The stack trace we saw was pointing to undinci
https://github.com/nodejs/undici/issues/583
Which looks is the nodejs fetch implementatin. So we try to use
node-fetch to see if this mitigates the error