feat(schema): migrate documentLogId pointer from evaluationResults to providerLogs (#536)
Added new fields `evaluatedProviderLogId` and `evaluationProviderLogId`
to the `evaluationResults` schema. These fields are intended to replace
the `documentLogId` and providerLogId fields. The reason is that we want
to track the precise provider log that was evaluated, not the document
log. Because an evaluation can be run at different points in time over a
document and the document might have different providerlogs associated
with it at each point.