Allow caching system, user and assistant messages for Anthropic (#524)
* Allow caching system, user and assistant messages for Anthropic
We want to allow with our syntaxt for users to tag messages with
Anthropic cache control tag.

Also as part of this PR we extract first system messages in an anthropic
prompt as part of the configuration. This way we can remove the warning

And last change in this PR is that we allow system messages to have more
than one message. This is something that's not valid in OpenAI but is
valid in Anthropic so we allow this at compiler level but we restric for
OpenAI at Rules level

* chore: updated docs

* chore: update docs 2

* chore: updated docs 3

* chore: updated docs 4

---------

Co-authored-by: Gerard Clos <clos.gerard@gmail.com>