feat(core): add SQL migration for updating evaluated provider log IDs
This commit introduces a new SQL migration file `0097_mushy_xavin.sql`
that updates the `evaluated_provider_log_id` in the `evaluation_results`
table. The migration ensures that the latest provider log is associated
with each evaluation result, provided the `evaluated_provider_log_id` is
currently null. This allows us to point to the actual provider log that
was evaluated, instead of the document log which does not represent a
particular conversation in time.