Add dataset ID relation to document versions (#570)
We want to allow users to pick parameters for a document from the rows
of a dataset they have in their workspace. We link that dataset_id with
the document version.
The `dataset_id` is an attribute that count as content change to be
consired when a new version of the document is merged. We do this to
allow users iterate the prompt through time using different datasets. So
they can back to the pass and see the old prompt working with the old
dataset.