Data schema for public prompts. (#664)
The idea is to allow users in prompt UI to publish the prompt so people
on their organization without access to Latitude can interact with this
prompt.

I'm thinking on doing a different table with published documents. I'm
doing this way so we can share public UUIDs and not internal document
uuid. Also because we want to add extra configuration like if the user
want to allow to consumers of this public prompt to follow up the
conversation with the AI.