fix(telemetry): concurrency fixes
In theory open telemetry spans are unique per trace but the reality is
that nothing stops clients from sending spans with the same ID multiple
times, and indeed they do. We handle this case by merging the spans on
conflict where it is sensible to do so.

Also implemented some minor UI fixes and type fixes.