feat: fix workers/queues connection config
We were initializing queues and workers with enableOfflineQueue: true
but that is not the correct production configuration. Queues need to be
initialized with enableOfflineQueue: true but workers need to be initialized
with enableOfflineQueue: false.

This commit fixes the issue by initializing queues with enableOfflineQueue:
true and workers with enableOfflineQueue: false.

I've also simplified some code in the queues initialization.