feature: autoscale integrations
We automatically scale mcp servers down to 0 replicas if they are not
used in the last 10 minutes in free plans. We automatically scale them
up again if they are requested in a prompt run.

Also refactored code to reuse mcp connections during a prompt stream in
order to decrease connection overhead.