Fix race conditions in the way we load dataset parameters on playground (#1040)
We did in a way that data is not persisted also in local storage and
this makes all the UI behave jumpy. Now we persist also in localStorage
as with old datasets. Also tried to do an implementation where datasetV2
and datasetRows are not fetched until user clicks on dataset parameters
but this provokes raise condition document emiting metadata parameters
and datasets not being set correctly