feat: improve count logs query performance
This query was unnecessarily slow by joining to multiple tables to check
for tenancy when we were already passing an identifier to the query.

Even if all joins used indeces the fact that the db had to join the
table to do the counting was expensive due to the size of the table
(theory, I don't have data to back this up).