Improve API and SDK error delivering (#1333)
While working on SDK examples I was using our SDK.run method in
non-streaming method. I forgot to add one of the parameters used in the
prompt to the parameters in the SDK and I noticed that we are returning
a generic 500 API error. We should return to the user of the SDK the
error that happened during the run of the prompt.