Fix opening evaluation delete modal leaving the UI blocked
The problem was that the dropdown remains open and when the modal close
the `pointer-events: none` in the body that radix dropdown adds appears
again and is not possible to click anything. The fix is to close the
dropdown when after an item in the dropdown is clicked.

Fix this issue:
https://github.com/latitude-dev/latitude-llm/issues/1362