feat: New prompt run engine
This commit refactors stream management in a pretty significant way.

** Removed old compiler **
We've removed the old compiler and any references to promptLVersion
checking – as of June 1st we no longer support the old syntax.

** Removed pause/resume prompt logic **
There's no difference between a tool that requires input and a tool that
calls an async method (such as WebSearch). As such, tools that require
user/async input no longer pause the prompt. Instead we just wait with
the stream open and a timeout until the user/async process responds via
a webhook endpoint and pubsub event.

** Removed runAgent and most of the agent-specific code **
Agents are now just prompts that can execute more than one step and
respond with text when all other tool calls have been resolved. This
makes it so that most differentitation between agents and regular chains
is now gone and we've deleted most agent-specific code.

** Refactor stream management **
Stream management was overly complex. We now implement a stream
management strategy with specific process executors: one for chains and
one for chat messages. This ensures a common abstraction to handle
streams from llm providers while giving us the flexibility to call the
llm provider with different parameters for different use cases.