Avoid having NextJS client router side cache when page is dynamic
In a previous commit we did the server side page dynamic
https://github.com/latitude-dev/latitude-llm/commit/80339803ae86550730fcd33bb3aff46c93a06372

This made the server rendering part of the Next.js process not to be
cached. But nextjs has a client side routing caching mechanish where
props for the RSC (React Server Components) are cached for a time.

I think we're seing this cache when we don't see the document.content
not being updated.

Really I have no idea. Just want to try.

More info here for `stateTimes`:
https://nextjs.org/docs/app/api-reference/config/next-config-js/staleTimes

And here for Client-Side router cache:
https://nextjs.org/docs/app/guides/caching#client-side-router-cache