Fix document editor cache issue (#1568)
* Revert "bugfix(Editor) - remove cached documents"

This reverts commit 92a6dace8451f7cefba680d93cc5e05960fe0f90.

* Fix document editor cache issue.
The editors get the value from the `useState` inside `useDocumentValueContext`.
This state is set from the props comming from the backend. Somehow editing a document, going other page and comming back is showing
stale text in the text editor. Now we get that info fresh from the `swr`
store for the old and new editors