tmp: try to fix nextjs cache issue
I noticed the client side component is actually rendered in backend but
only the first time, whilst subsequent navigations do not trigger this
backend rendering — the nextjs documentation confirmed as much. I have
the feeling that adding a key to the client side component will force a
remounting of the component on each navigation.