Bring back to our code mailgun nodemailer adapter because is not maintained
We added an import in our apps/web next.js app and it breaks the Next.JS
build because this import from `packages/core` imports our mailers code
which use https://www.npmjs.com/package/nodemailer-mailgun-transport
which is super old and use `babel-core` which is deprecated.

The lib is just a file so moving it to our core and updating their
dependencies.