bugfix(experiments) - Add Metadata Provider to experiments page
With the documentEditor design, we migrated the metadata web worker logic to a provider used within the documentEditor (old and new). However, we didnt change the experiments page component which used to call this web worker logic hook. It worked well now as the metadata is stored in a zustand store, so when going from the document editor to the experiments page, it worked well, but if we refreshed within the experiments page, the metadata was lost.