Add `userMessage` to `run` method of the Typescript SDK in beta version (#1592)
In the beta version of the SDK v5 we want to allow to run a prompt with
a user message on it. We want this to allow prompts to be run like if in
GPT where you can start asking something to a preconfigured AI. In this
case could be a Latitude AI agent.