feat: migrate provider log data to object storage (#1609)
Provider logs is one of our largest tables and the size per row is quite
substantial given we store the full convo and related artifacts in the
table.

This commit moves non aggregagable data (mainly json columns) to
object storage and updates the code to account for this fact,
maintaining backwards compat.