Update trigger list when a trigger is created or deleted (#1623)
* Update trigger list when a trigger is created or deleted

* Fix select trigger not showing as disabled
I moved the trigger primitive up in the element here:
https://github.com/latitude-dev/latitude-llm/pull/1541

That is a problem because disabled styles `disabled:opacity-50` only
apply on buttons and similar elements not in `div`. What I did to fix it
is remove the inner div and leave the radix trigger primitive use a
button. So disable styles work again