Refactor Latte Integration Tools (#1648)
We have too many tools with similar functionalities for finding integration, tools and triggers. This PR simplifies this situation.

We will now have the following tools:

SearchAvailableIntegrations - Search apps of pipedream using a query
ListExistingIntegrations - List existing integrations of the workspace
ListIntegrationTools - List the tools of an integration (existing in the workspace or not)
ListIntegrationTriggers - List the triggers of an integration (existing in the workspace or not)