Remove redundant prevContext
We didnt understand the use of prevContext when we implemented it, and we were using it incorrectly. Its used for pagination purposes, and as we were always sending it, at times it would make certain selects break when there was no more pages. We're removing it as we don't implement pagination yet with pipedream components