Implement pro and new team subscriptions. (#1669)
We want to allow users to buy new subscriptions while we granfather the
old team V1 subscription. In this changes we allow users to go to the
pricing page if they're in the free plan and choose between pro and team
v2 plans. If they're already in pro plan they can go to the customer
portal the do the upgrade. We need to handle the new subscriptions and
also the upgrades on the Stripe subscriptions webhook