Vercel sdk v5 maybe this time (#1766)
* Vercel SDK v5
This PR introduce Vercel SDK v5 which force us also to upgrade Zod from
v3 to v4 and also to change our server action lib because is not
compatible with Zod 4.

* Fix unhandle Vercel SDK error NoOutputGeneratedError
The way we handle this is by listening to `onError` callback on
`streamText` and throw the error before we start reading from the
streamText promises which will fail anyway because when Vercel gets
provider API error they return the NoOutputGeneratedError content which
is not handle in our side. Throwing on their onError callback we avoid
this problem