Adapt new onboarding with deletion chat triggers (#1806)
With the deletion of chat triggers, we needed to make some changes for the onboarding to make sense.

From now on, there will always be a chat box to talk to your agent within the trigger agent page (as we have now with the new home page). Regardless if you have a trigger or not, you will now be able to "trigger" your agent with the chat.

In addition to this, the simulate with AI feature was broken, now its fixed