Fix validation error issue in dataset modal
The first letter of the errors was appearing when a validation error occurred when creating datasets. This was due to how we were flattening zod validation errors in the API, where we returned a key value record, where the value were strings instead of array of strings. We could have left it returning a string and modify frontend logic, however as our input atoms ask for an array of strings, and zod can return multiple validation errors for a single attribute, its more extendable to modifying the flattening logic to return an array of strings as a value, as we can reuse this logic for more forms.