Remove duplicate tool calls when assistant message included tool call (#1840)
When we were calling tool calls, we would merge the assistant content with the toolCalls argument, regardless if the content of the assistant message had the tool call type already there. This resulted in an assistant message with two tool calls within the content with the same id. Some providers don't mind this, but Anthropic throws an error, which is the one Will was facing with Latte. This fix fixes that