Add cost estimation to APIv3 runs (#1905)
In this PR, added the cost argument to our API calls of prompts.run, conversations.chat, and conversation.attach so the user can see an average cost estimation of their runs.

This is only available if the isStream is false for now and doesnt count the cost of the subagents for that document. However, its a good quick win.