Change issues <-> evaluation results relation to be HAS_MANY
Now an evaluation results is pointing to an issue. But this is a problem
because when an issue is merged we dont want  to lose the annotations
that were made in it. So an evaluation result will point now to MANY
issues. The original that was merged and the new issue where the
original was merged

WIP