Aligned cleanup grace period between Redis and BullMQ + new maintenance cleanup job of orphan background runs
We were getting a few conversations not found errors when users tried to attach to a conversation that already ended. However, we had a missalignment between the cleanup time of the Redis stream and the bullMQ job, where the Redis stream takes 10s to remove, while BullMQ removes the job instantaniously once completed. This PR aligns these times, and additionally adds a maintenance job to cleanup orphan background jobs that remain in the Redis cache due to possible edge cases.