Migrate background jobs proactively  (#1920)
We're adding the migration of JSON to HASH values within every Redis action, potentially hanging up the job.

This change makes the migration only when needed, if an error of WRONGTYPE is thrown, and also added a maintenance job to run to migrate all of the active runs to HASH format.