Fix double count in runs page (#1918)
We had an issue where if you were in the runs page and you ran documents from somewhere else, the counter would count twice each run. This was because we had the sidebar and the runs page using the useActiveRunsCount hook, which would create two listeners of the WS event, and mutate twice the frontend state. This fix fixes that by creating a provider that uses this hook, and instantiating it in the layout for the children to use, and avoid duplicate counts.