Align lock duration to backgroundJob potential timeout
We have a grace period of streaming processing in the backgroundJob of 10 minutes, but were using the default BullMQ lock duration of 30 seconds. So if a backgroundJob takes 5 minutes, after 30 it gets considered stalled in BullMQ and moves to waiting for another worker to get it. In the meantime, its still executing, and if another worker later on gets it, it will start executing the job again, creating potential duplicates. To solve this, we can match the lock duration of the runs queue to the grace period of streaming, plus an additional buffer for post processing (if its an experiment, etc.).