Persists isEscalating state in DB so we can implement alerts
We want to implement a simple alerting system so users can be notified
when an issue starts escalating. One way of doing is to check if the
issue is escalating when an increment in histogram occours (or
decrement). If is escalating we sent an email to the users on the
repository to notify of the issue. But we dont want to notify again for
the same issue if the issue is already escalating. To do this we
persists escalating_at so we can skip the sending of the email