Fix refiner and document suggestor
These two features were still using the old provider log mechanism to find the documentLog. Therefore, for the newer logs created, the evaluatedLogId wasn't found and it would break. This fix fixes this by adding the new method using spans and traces, and falls back to the evaluatedLogId if they're not found (for old logs).