Add flag for big accounts
We want to avoid overloading the system with big accounts. In those case
with auxiliary workflows like sending a weekly email with data on the
workspace we are adding the ability of skiping this queries for
workspaces marked as too big. Until we improve the platform to support
any kind of scale this is a scape hatch.